tests: remove dependency on snapcraft for integration tests
The integration tests import snapcraft into the suite, this causes all
sorts of problems, the most concerning one is test independence.
