+++ spread.yaml
@@ -259,5 +259,9 @@ suites:
      apt-get install --yes gcc g++ make python3-dev python3-pip python3-wheel libffi-dev libsodium-dev libapt-pkg-dev squashfs-tools xdelta3 bzr git mercurial subversion
      pip3 install --user --upgrade pip
      $PIP_COMMAND install --user -r /snapcraft/requirements.txt -r /snapcraft/requirements-devel.txt
+     # Move the snapcraft modules out of the way
+     mv /snapcraft/snapcraft /snapcraft.bak
+   restore: |
+     mv /snapcraft.bak /snapcraft/snapcraft
 
 path: /snapcraft/
