+++ tests/integration/general/test_architectures.py
@@ -16,9 +16,9 @@
 
 import os
 
+import yaml as yaml_utils
 from testtools.matchers import Equals
 
-from snapcraft import yaml_utils
 from tests import integration
