+++ tests/integration/general/test_snapcraftctl_set_grade.py
@@ -18,9 +18,9 @@ import os
 import subprocess
 import textwrap
 
+import yaml as yaml_utils
 from testtools.matchers import Equals, Contains
 
-from snapcraft import yaml_utils
 from tests import integration
