+++ tests/integration/lifecycle/test_build_properties.py
@@ -16,9 +16,9 @@
 
 import os
 
+import yaml as yaml_utils
 from testtools.matchers import Equals, FileExists
 
-from snapcraft import yaml_utils
 from tests import integration
