+++ tests/integration/plugins_nodejs/test_nodejs_plugin.py
@@ -16,9 +16,9 @@
 import os
 
 import testscenarios
+import yaml as yaml_utils
 from testtools.matchers import FileExists
 
-from snapcraft import yaml_utils
 from tests import integration
