+++ tools/travis/run_integration_test.sh
@@ -76,6 +76,7 @@ $lxc file push --recursive "$project_path" test-runner/root/
 $lxc exec test-runner -- sh -c "apt install -y bzr git libnacl-dev libssl-dev libsodium-dev libffi-dev libapt-pkg-dev mercurial python3-pip subversion sudo snapd"
 $lxc exec test-runner -- sh -c "python3 -m pip install -I pip==9.0.3"
 $lxc exec test-runner -- sh -c "python3 -m pip install -r snapcraft/requirements-devel.txt -r snapcraft/requirements.txt"
+$lxc exec test-runner -- sh -c "rm -rf snapcraft/snapcraft"
 $lxc exec test-runner -- sh -c "${SNAPCRAFT_INSTALL_COMMAND:-sudo snap install snapcraft-pr$TRAVIS_PULL_REQUEST.snap --dangerous --classic}"
 $lxc exec test-runner -- sh -c "cd snapcraft && ./runtests.sh $test_suite"
